Move fixture classes to Fixture namespace
Also rename the strategy class to better match how it operates.
