+++ tests/TestCase/EmbeddedDocumentTest.php
@@ -19,7 +19,7 @@ namespace Cake\ElasticSearch\Test\TestCase;
 use Cake\Core\Configure;
 use Cake\Datasource\ConnectionManager;
 use Cake\ElasticSearch\Index;
-use Cake\ElasticSearch\TestSuite\ElasticStateReset;
+use Cake\ElasticSearch\TestSuite\Fixture\DeleteQueryStrategy;
 use Cake\TestSuite\TestCase;
 
 /**
@@ -27,7 +27,7 @@ use Cake\TestSuite\TestCase;
  */
 class EmbeddedDocumentTest extends TestCase
 {
-    protected $stateResetStrategy = ElasticStateReset::class;
+    protected $stateResetStrategy = DeleteQueryStrategy::class;
 
     public $fixtures = ['plugin.Cake/ElasticSearch.Profiles'];
