+++ tests/TestCase/IndexTest.php
@@ -21,7 +21,7 @@ use Cake\Datasource\EntityInterface;
 use Cake\ElasticSearch\Document;
 use Cake\ElasticSearch\Index;
 use Cake\ElasticSearch\IndexRegistry;
-use Cake\ElasticSearch\TestSuite\ElasticStateReset;
+use Cake\ElasticSearch\TestSuite\Fixture\DeleteQueryStrategy;
 use Cake\Event\EventInterface;
 use Cake\TestSuite\TestCase;
 use Elastica\Exception\NotFoundException;
@@ -31,7 +31,7 @@ use Elastica\Exception\NotFoundException;
  */
 class IndexTest extends TestCase
 {
-    protected $stateResetStrategy = ElasticStateReset::class;
+    protected $stateResetStrategy = DeleteQueryStrategy::class;
 
     public $fixtures = ['plugin.Cake/ElasticSearch.Articles'];
