+++ tests/TestCase/View/Form/DocumentContextTest.php
@@ -23,7 +23,7 @@ use Cake\Datasource\ConnectionManager;
 use Cake\ElasticSearch\Document;
 use Cake\ElasticSearch\Index;
 use Cake\ElasticSearch\IndexRegistry;
-use Cake\ElasticSearch\TestSuite\ElasticStateReset;
+use Cake\ElasticSearch\TestSuite\Fixture\DeleteQueryStrategy;
 use Cake\ElasticSearch\View\Form\DocumentContext;
 use Cake\Http\ServerRequest;
 use Cake\TestSuite\TestCase;
@@ -35,7 +35,7 @@ use TestApp\Model\Document\Article;
  */
 class DocumentContextTest extends TestCase
 {
-    protected $stateResetStrategy = ElasticStateReset::class;
+    protected $stateResetStrategy = DeleteQueryStrategy::class;
 
     /**
      * Fixtures to use.
