+++ tests/bootstrap.php
@@ -31,7 +31,7 @@ define('LOGS', TMP . 'logs' . DS);
 use Cake\Cache\Cache;
 use Cake\Core\Configure;
 use Cake\Datasource\ConnectionManager;
-use Cake\ElasticSearch\TestSuite\MappingGenerator;
+use Cake\ElasticSearch\TestSuite\Fixture\MappingGenerator;
 use Cake\Routing\Router;
 
 Configure::write('debug', true);
